@@ -32,29 +32,29 @@
   "dependencies": {
     "@amcharts/amcharts4": "^4.10.22",
     "@amcharts/amcharts4-geodata": "^4.1.21",
-    "@angular/animations": "^14.2.0",
-    "@angular/cdk": "^14.2.0",
-    "@angular/common": "^14.2.0",
-    "@angular/compiler": "^14.2.0",
-    "@angular/core": "^14.2.0",
-    "@angular/forms": "^14.2.0",
-    "@angular/localize": "^14.2.0",
-    "@angular/material": "^14.2.0",
-    "@angular/platform-browser": "^14.2.0",
-    "@angular/platform-browser-dynamic": "^14.2.0",
-    "@angular/router": "^14.2.0",
-    "@europeana/metis-ui-maintenance-utils": "0.0.4",
-    "ngx-matomo": "^1.1.0",
+    "@angular/animations": "^16.2.6",
+    "@angular/cdk": "^16.2.5",
+    "@angular/common": "^16.2.6",
+    "@angular/compiler": "^16.2.6",
+    "@angular/core": "^16.2.6",
+    "@angular/forms": "^16.2.6",
+    "@angular/localize": "^16.2.6",
+    "@angular/material": "^16.2.5",
+    "@angular/platform-browser": "^16.2.6",
+    "@angular/platform-browser-dynamic": "^16.2.6",
+    "@angular/router": "^16.2.6",
+    "@europeana/metis-ui-maintenance-utils": "0.0.5",
+    "ngx-matomo": "^2.0.0",
     "pdfmake": "^0.1.72",
     "rxjs": "^7.8.1",
     "ts-loader": "^9.3.1",
     "tslib": "^2.0.0",
-    "zone.js": "~0.11.4"
+    "zone.js": "^0.13.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^14.2.1",
-    "@angular/cli": "^14.2.1",
-    "@angular/compiler-cli": "^14.2.0",
+    "@angular-devkit/build-angular": "^16.2.3",
+    "@angular/cli": "^16.2.3",
+    "@angular/compiler-cli": "^16.2.6",
     "@cypress/webpack-preprocessor": "^5.12.2",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "^2.0.10",
@@ -65,11 +65,10 @@
     "@typescript-eslint/types": "^5.36.0",
     "axe-core": "^4.4.1",
     "codelyzer": "^6.0.0",
-    "cypress": "^10.5.0",
-    "cypress-axe": "^0.14.0",
+    "cypress": "^10.11.0",
     "eslint": "^8.23.0",
-    "eslint-plugin-rxjs": "^5.0.2",
-    "jasmine": "^3.9.0",
+    "eslint-plugin-rxjs": "^5.0.3",
+    "jasmine": "^5.1.0",
     "jasmine-core": "^3.6.0",
     "jasmine-spec-reporter": "~5.0.0",
     "karma": "^6.4.0",
@@ -81,7 +80,7 @@
     "npm-run-all": "^4.1.5",
     "prettier": "^2.4.1",
     "ts-node": "~8.3.0",
-    "typescript": "^4.7.4",
+    "typescript": "^5.1.6",
     "typescript-eslint": "0.0.1-alpha.0"
   }
 }
